refactor: split requirements.txt collector as separate module
