+++ test/requirements.txt.collector.test.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { ReqDependencyCollector } from '../src/collector';
+import { ReqDependencyCollector } from '../src/requirements.txt.collector';
 
 describe('PyPi requirements.txt parser test', () => {
     const collector:ReqDependencyCollector = new ReqDependencyCollector();
@@ -10,23 +10,19 @@ describe('PyPi requirements.txt parser test', () => {
             c>=10.1
             d<=20.1.2.3.4.5.6.7.8
         `);
-        expect(deps.length).equal(4);
-        expect(deps[0]).is.eql({
+        expect(deps).is.eql([{
           name: {value: 'a', position: {line: 0, column: 0}},
           version: {value: '1', position: {line: 1, column: 4}}
-        });
-        expect(deps[1]).is.eql({
+        },{
           name: {value: 'b', position: {line: 0, column: 0}},
           version: {value: '2.1.1', position: {line: 2, column: 16}}
-        });
-        expect(deps[2]).is.eql({
+        },{
           name: {value: 'c', position: {line: 0, column: 0}},
           version: {value: '10.1', position: {line: 3, column: 16}}
-        });
-        expect(deps[3]).is.eql({
+        },{
           name: {value: 'd', position: {line: 0, column: 0}},
           version: {value: '20.1.2.3.4.5.6.7.8', position: {line: 4, column: 16}}
-        });
+        }]);
     });
 
     it('tests requirements.txt with comments', async () => {
@@ -37,19 +33,16 @@ describe('PyPi requirements.txt parser test', () => {
             d<=20.1.2.3.4.5.6.7.8
             # done
         `);
-        expect(deps.length).equal(3);
-        expect(deps[0]).is.eql({
+        expect(deps).is.eql([{
           name: {value: 'a', position: {line: 0, column: 0}},
           version: {value: '1', position: {line: 2, column: 16}}
-        });
-        expect(deps[1]).is.eql({
+        },{
           name: {value: 'c', position: {line: 0, column: 0}},
           version: {value: '', position: {line: 4, column: 1}} // column shouldn't matter for empty versions
-        });
-        expect(deps[2]).is.eql({
+        },{
           name: {value: 'd', position: {line: 0, column: 0}},
           version: {value: '20.1.2.3.4.5.6.7.8', position: {line: 5, column: 16}}
-        });
+        }]);
     });
 
     it('tests empty lines', async () => {
@@ -58,11 +51,10 @@ describe('PyPi requirements.txt parser test', () => {
             a==1
 
         `);
-        expect(deps.length).equal(1);
-        expect(deps[0]).is.eql({
+        expect(deps).is.eql([{
           name: {value: 'a', position: {line: 0, column: 0}},
           version: {value: '1', position: {line: 3, column: 16}}
-        });
+        }]);
     });
 
     it('tests deps with spaces before and after comparators', async () => {
@@ -72,14 +64,12 @@ describe('PyPi requirements.txt parser test', () => {
                   b        <=     10.1               
 
         `);
-        expect(deps.length).equal(2);
-        expect(deps[0]).is.eql({
+        expect(deps).is.eql([{
           name: {value: 'a', position: {line: 0, column: 0}},
           version: {value: '1', position: {line: 2, column: 24}}
-        });
-        expect(deps[1]).is.eql({
+        },{
           name: {value: 'b', position: {line: 0, column: 0}},
           version: {value: '10.1', position: {line: 4, column: 35}}
-        });
+        }]);
     });
 });
